+++ src/contracts/index.ts
@@ -4,6 +4,7 @@ export type { ErrorConstructorType } from './ErrorConstructorType.js';
 export type { ExecuteOptionsInterface } from './ExecuteOptionsInterface.js';
 export type { NodeInterface } from './NodeInterface.js';
 export type { OperationContract } from './OperationContract.js';
+export type { OperationContractFragment } from './OperationContractFragment.js';
 export type { RetryPolicyOptionsInterface } from './RetryPolicyOptionsInterface.js';
 export type { SchedulerHandle } from './SchedulerHandle.js';
 export type { SchedulerProvider } from './SchedulerProvider.js';